// frontend/ClientApp/src/types/user-role.ts
export type UserRole =
  | 'TEMPLATE_CREATOR'
  | 'TEMPLATE_REVIEWER'
  | 'TEMPLATE_APPROVER'
  | 'TEMPLATE_MANAGER'
  | 'CONTRACT_CREATOR'
  | 'CONTRACT_REVIEWER'
  | 'CONTRACT_APPROVER'
  | 'CONTRACT_MANAGER'
  | 'CONTRACT_SIGNER'
  | 'CONTRACT_OBSERVER'
  | 'ARCHIVE_MANAGER'
  | 'AUDITOR'
  | 'SYSTEM_ADMINISTRATOR'
  | 'COMPLIANCE_OFFICER'
  | 'INTEGRATION_MANAGER'
  | 'PROCESS_ORCHESTRATOR'
  | 'VALIDATOR'